+++ src/services/team.service.js
@@ -2,17 +2,17 @@ import api from "@/lib/api";
 
 export const TeamService = {
   getTeams: async () => {
-    const response = await api.get("/team");
+    const response = await api.get("/public/team");
     return response.data;
   },
 
   getTeamMemberById: async (id) => {
-    const response = await api.get(`/team/${id}`);
+    const response = await api.get(`/admin/team/${id}`);
     return response.data;
   },
 
   createTeamMember: async (formData) => {
-    const response = await api.post("/team", formData, {
+    const response = await api.post("/admin/team", formData, {
       headers: {
         "Content-Type": "multipart/form-data",
       },
@@ -21,7 +21,7 @@ export const TeamService = {
   },
 
   updateTeamMember: async (id, formData) => {
-    const response = await api.put(`/team/${id}`, formData, {
+    const response = await api.put(`/admin/team/${id}`, formData, {
       headers: {
         "Content-Type": "multipart/form-data",
       },
@@ -30,7 +30,7 @@ export const TeamService = {
   },
 
   deleteTeamMember: async (id) => {
-    const response = await api.delete(`/team/${id}`);
+    const response = await api.delete(`/admin/team/${id}`);
     return response.data;
   },
 };
